Move page feedback interactor logic to superclass
