+++ lib/core.rb
@@ -54,6 +54,7 @@ module Core
   autoload :NewsReader, 'core/interactor/news_reader'
   autoload :PageFeedbackCreator, 'core/interactor/page_feedback_creator'
   autoload :PageFeedbackUpdator, 'core/interactor/page_feedback_updator'
+  autoload :PageFeedbackAction, 'core/interactor/page_feedback_action'
   autoload :Searcher, 'core/interactor/searcher'
   autoload :StaticPageReader, 'core/interactor/static_page_reader'
   autoload :VideoReader, 'core/interactor/video_reader'
