+++ lib/core/interactor/page_feedback_creator.rb
@@ -1,18 +1,7 @@
 module Core
-  class PageFeedbackCreator
+  class PageFeedbackCreator < PageFeedbackAction
     def call(params)
-      data = repository.create(params)
-      return false if data.blank?
-
-      entity.new(data['id'], data)
-    end
-
-    def repository
-      Core::Registry::Repository[:page_feedback]
-    end
-
-    def entity
-      Core::PageFeedback
+      action(:create, params)
     end
   end
 end
